Kapo_Back/requirements.txt
pip>=19.3.1
Django>=3.0.1
djangorestframework>=3.11.0
asgiref>=3.2.3
pytz>=2019.3
setuptools>=42.0.2
sqlparse>=0.3.0
django-phonenumber-field>=4.0.0
phonenumbers>=8.11.1
Pillow>=6.2.1
python-gettext>=4.0
gunicorn==19.9.0
dj-database-url==0.5.0
django-cors-headers>=3.2.0
djangorestframework-jwt>=1.11.0